backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
numpy
websockets
openai
git+https://github.com/TheStageAI/TheWhisper.git
